+++ Pool/Analytics/AnalyticsViewController.swift
@@ -1,5 +1,5 @@
 //
-//  ScoresViewController.swift
+//  AnalyticsViewController.swift
 //  Pool
 //
 //  Created by mpc on 10/1/21.
@@ -7,12 +7,12 @@
 
 import UIKit
 
-class ScoresViewController: UIViewController {
+class AnalyticsViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        // Do any additional setup after loading the view.
+        view.backgroundColor = .cyan
     }
