+++ Pool/Home/HomeViewController.swift
@@ -1,0 +1,17 @@
+//
+//  HomeViewController.swift
+//  Pool
+//
+//  Created by mpc on 10/1/21.
+//
+
+import UIKit
+
+class HomeViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .brown
+    }
+}
