--- Pool/HomeViewController.swift
@@ -1,29 +1,0 @@
-//
-//  HomeViewController.swift
-//  Pool
-//
-//  Created by mpc on 10/1/21.
-//
-
-import UIKit
-
-class HomeViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
