+++ Pool/Scores/ScoresView.swift
@@ -1,0 +1,20 @@
+//
+//  ScoresView.swift
+//  Pool
+//
+//  Created by mpc on 10/1/21.
+//
+
+import UIKit
+
+class ScoresView: UIView {
+
+    /*
+    // Only override draw() if you perform custom drawing.
+    // An empty implementation adversely affects performance during animation.
+    override func draw(_ rect: CGRect) {
+        // Drawing code
+    }
+    */
+
+}
